earthquake_loader/requirements.txt
defopt==6.0.1
pandas==1.1.3
pause==0.2
kafka-python==2.02
openeew==0.5.0
aioboto3==7.0.0
aiobotocore==0.12.0
boto3==1.12.15
botocore==1.15.15
# nest_asyncio